[#] Large chunk with bug fix style and refactor.

+++ src/kern/pipe.c
@@ -9,12 +9,13 @@
 
 /* ----------------------------------------------------------------------- */
 /**  */
-kPipe_t * fs_create_pipe(kInode_t *ino)
+kPipe_t *fs_create_pipe(kInode_t *ino)
 {
   int vmaRg = VMA_FIFO | VMA_READ | VMA_WRITE;
   kPipe_t *pipe;
 
   klock(&ino->lock_);
+
   if (ino->pipe_) {
     kunlock(&ino->lock_);
     return ino->pipe_;
@@ -25,6 +26,7 @@ kPipe_t * fs_create_pipe(kInode_t *ino)
 
   if (ino->stat_.length_ == 0)
     ino->stat_.length_ = PAGE_SIZE;
+
   pipe->size_ = ALIGN_UP(ino->stat_.length_, PAGE_SIZE);
   klock(&kSYS.mspace_->lock_);
   pipe->mmap_ = area_map(kSYS.mspace_, pipe->size_, vmaRg);
@@ -53,9 +55,10 @@ static size_t fs_pipe_newline(kPipe_t *pipe)
 {
   size_t i, cap;
   size_t max = pipe->size_ - pipe->rpen_;
-  char* address = (char*)(pipe->mmap_->address_ + pipe->rpen_);
+  char *address = (char *)(pipe->mmap_->address_ + pipe->rpen_);
 
   cap = MIN(max, pipe->avail_);
+
   for (i = 0; i < cap; ++i) {
     if (address[i] == '\n')
       return i + 1;
@@ -64,8 +67,9 @@ static size_t fs_pipe_newline(kPipe_t *pipe)
   if (max > pipe->avail_)
     return 0;
 
-  address = (char*)(pipe->mmap_->address_);
+  address = (char *)(pipe->mmap_->address_);
   cap = MAX(0, pipe->avail_ - max);
+
   for (i = 0; i < cap; ++i) {
     if (address[i] == '\n')
       return i + max + 1;
@@ -76,11 +80,11 @@ static size_t fs_pipe_newline(kPipe_t *pipe)
 
 /* ----------------------------------------------------------------------- */
 /**  */
-int fs_pipe_read(kInode_t *ino, void* buf, size_t lg)
+ssize_t fs_pipe_read(kInode_t *ino, void *buf, size_t lg)
 {
-  size_t bytes = 0;
+  ssize_t bytes = 0;
   size_t cap = 0;
-  void* address;
+  void *address;
   kPipe_t *pipe = ino->pipe_;
 
   assert (S_ISFIFO(ino->stat_.mode_) || S_ISCHR(ino->stat_.mode_));
@@ -91,23 +95,28 @@ int fs_pipe_read(kInode_t *ino, void* buf, size_t lg)
 
   /// @todo Mutex on pipes
   mtx_lock(&pipe->mutex_);
+
   while (lg > 0) {
     if (pipe->rpen_ >= pipe->size_)
       pipe->rpen_ = 0;
 
     /* Get Address */
-    address = (void*)(pipe->mmap_->address_ + pipe->rpen_);
+    address = (void *)(pipe->mmap_->address_ + pipe->rpen_);
 
     /* Capacity ahead */
     cap = pipe->size_ - pipe->rpen_;
+
     if (pipe->flags_ & FP_BY_LINE)
       cap = MIN(cap, fs_pipe_newline(pipe));
     else
       cap = MIN(cap, pipe->avail_);
+
     cap = MIN(cap, lg);
+
     if (cap == 0) {
       if (!(pipe->flags_ & FP_BLOCK) || bytes != 0)
         break;
+
       wait_for(&pipe->mutex_, WT_PIPE_READ, &pipe->waiting_);
       continue;
     }
@@ -119,6 +128,7 @@ int fs_pipe_read(kInode_t *ino, void* buf, size_t lg)
     pipe->avail_ -= cap;
     bytes += cap;
     buf = ((char *)buf) + cap;
+
     if (pipe->flags_ & FP_BY_LINE && ((char *)buf)[-1] == '\n')
       break;
   }
@@ -130,23 +140,23 @@ int fs_pipe_read(kInode_t *ino, void* buf, size_t lg)
 
 /* ----------------------------------------------------------------------- */
 /**  */
-size_t fs_pipe_write(kInode_t *ino, const void* buf, size_t lg)
+ssize_t fs_pipe_write(kInode_t *ino, const void *buf, size_t lg)
 {
   size_t i;
-  size_t bytes = 0;
+  ssize_t bytes = 0;
   size_t cap = 0;
-  void* address;
+  void *address;
   kPipe_t *pipe = ino->pipe_;
-  kWait_t * wait;
-  kWait_t * iter;
+  kWait_t *wait;
+  kWait_t *iter;
   bool haveNl = false;
 
   assert (S_ISFIFO(ino->stat_.mode_) || S_ISCHR(ino->stat_.mode_));
 
   /* Search for '\n' */
   if (pipe->flags_ & FP_BY_LINE) {
-    for (i=0; i<lg; ++i)
-      if (((char*)buf)[i] == '\n') {
+    for (i = 0; i < lg; ++i)
+      if (((char *)buf)[i] == '\n') {
         haveNl = true;
         break;
       }
@@ -157,6 +167,7 @@ size_t fs_pipe_write(kInode_t *ino, const void* buf, size_t lg)
     pipe = fs_create_pipe (ino);
 
   mtx_lock(&pipe->mutex_);
+
   if (pipe->flags_ & FP_WRITE_FULL) {
     if (pipe->size_ - pipe->avail_ < lg) {
       mtx_unlock(&pipe->mutex_);
@@ -169,12 +180,13 @@ size_t fs_pipe_write(kInode_t *ino, const void* buf, size_t lg)
       pipe->wpen_ = 0;
 
     /* Get Address */
-    address = (void*)(pipe->mmap_->address_ + pipe->wpen_);
+    address = (void *)(pipe->mmap_->address_ + pipe->wpen_);
 
     /* Capacity ahead */
     cap = pipe->size_ - pipe->wpen_;
     cap = MIN(cap, pipe->size_ - pipe->avail_);
     cap = MIN(cap, lg);
+
     if (cap == 0)
       break;
 
@@ -190,9 +202,11 @@ size_t fs_pipe_write(kInode_t *ino, const void* buf, size_t lg)
   // IF BLOCKED !
   if (haveNl) {
     iter = ll_first(&pipe->waiting_, kWait_t, lnd_);
+
     while (iter) {
       wait = iter;
       iter = ll_next(iter, kWait_t, lnd_);
+
       if (wait->reason_ == WT_PIPE_READ) {
         wait->reason_ = WT_HANDLED;
         sched_insert(kSYS.scheduler_, wait->thread_);
@@ -220,11 +234,12 @@ int fs_event(kInode_t *ino, int type, int value)
     return __seterrno(0);
   }
 
-  kprintf ("[E%x,%d-%s]", type,value,ino->name_);
+  kprintf ("[E%x,%d-%s]", type, value, ino->name_);
 
   // if (ino->subSystem_ == NULL) {
   if (fs_pipe_write(ino, &event, sizeof(event)) == 0)
     return __seterrno(EAGAIN);
+
   return __seterrno(0);
   // }
